requirements.txt
discord.py
zomboid-rcon
python-dotenv
coloredlogs
watchdog
paramiko